# backend/internal/extension/schema/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Amitia Extension Runtime API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/extensions/kernel/extension": {
      "get": {
        "description": "Get kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/extensions": {
      "get": {
        "description": "List kernel extensions",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "List kernel extensions",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/extensions/disable": {
      "post": {
        "description": "Disable kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Disable kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/extensions/enable": {
      "post": {
        "description": "Enable kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Enable kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
...
    "/extensions/kernel/extensions/pause": {
      "post": {
        "description": "Pause kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Pause kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
...
    "/extensions/kernel/extensions/resume-uninstall": {
      "post": {
        "description": "Resume kernel extension uninstall",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Resume kernel extension uninstall",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/extensions/rollback": {
      "post": {
        "description": "Rollback kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Rollback kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/extensions/uninstall": {
      "post": {
        "description": "Uninstall kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Uninstall kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/status": {
      "get": {
        "description": "Get kernel status",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get kernel status",
        "tags": [
          "Kernel"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/api"
    }
  ]
}